Restructured project for multiple routes
